factor 'legacy' pbh out into separate file
